driver-management-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6